refactor: remove unused scan helpers
